+++ torchao/sparsity/prototype/training/autograd.py
@@ -1,3 +1,14 @@
+import contextlib
+import ctypes
+import glob
+import warnings
+from functools import partial
+from pathlib import Path
+from typing import Any, Callable, Optional, Tuple, TypeVar, cast
+import torch
+from torch.sparse import SparseSemiStructuredTensor, SparseSemiStructuredTensorCUTLASS, SparseSemiStructuredTensorCUSPARSELT
+from collections.abc import Iterable
+from torch import nn
 
 if torch.__version__ >= "2.1.0":
     torch._dynamo.allow_in_graph(SparseSemiStructuredTensorCUSPARSELT)
